+++ client/src/components/layout/Sidebar.vue
@@ -116,7 +116,7 @@
     </div>
 
     <div class="p-4 border-t border-gray-800 text-sm text-gray-400">
-      <div>v0.40.1a-alpha</div>
+      <div>v0.40.2-alpha</div>
     </div>
   </aside>
 </template>